rename package flp (because shorter)
