+++ controllers/flp/flp_common_objects.go
@@ -1,4 +1,4 @@
-package flowlogspipeline
+package flp
 
 import (
 	"encoding/json"
