+++ controllers/flp/flp_ingest_objects.go
@@ -1,4 +1,4 @@
-package flowlogspipeline
+package flp
 
 import (
 	appsv1 "k8s.io/api/apps/v1"
