+++ controllers/flp/flp_ingest_reconciler.go
@@ -1,4 +1,4 @@
-package flowlogspipeline
+package flp
 
 import (
 	"context"
