+++ controllers/flp/flp_pipeline_builder.go
@@ -1,4 +1,4 @@
-package flowlogspipeline
+package flp
 
 import (
 	"fmt"
