+++ controllers/flp/flp_test.go
@@ -14,7 +14,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package flowlogspipeline
+package flp
 
 import (
 	"encoding/json"
