+++ controllers/flp/flp_transfo_reconciler.go
@@ -1,10 +1,11 @@
-package flowlogspipeline
+package flp
 
 import (
 	"context"
 
 	monitoringv1 "github.com/prometheus-operator/prometheus-operator/pkg/apis/monitoring/v1"
 	appsv1 "k8s.io/api/apps/v1"
+	ascv2 "k8s.io/api/autoscaling/v2"
 	corev1 "k8s.io/api/core/v1"
 	rbacv1 "k8s.io/api/rbac/v1"
 	"k8s.io/apimachinery/pkg/api/equality"
@@ -16,78 +17,78 @@ import (
 	"github.com/netobserv/network-observability-operator/pkg/helper"
 )
 
-// flpMonolithReconciler reconciles the current flowlogs-pipeline monolith state with the desired configuration
-type flpMonolithReconciler struct {
+// flpTransformerReconciler reconciles the current flowlogs-pipeline-transformer state with the desired configuration
+type flpTransformerReconciler struct {
 	*reconcilers.Instance
-	owned monolithOwnedObjects
+	owned transfoOwnedObjects
 }
 
-type monolithOwnedObjects struct {
-	daemonSet      *appsv1.DaemonSet
+type transfoOwnedObjects struct {
+	deployment     *appsv1.Deployment
 	promService    *corev1.Service
+	hpa            *ascv2.HorizontalPodAutoscaler
 	serviceAccount *corev1.ServiceAccount
 	configMap      *corev1.ConfigMap
-	roleBindingIn  *rbacv1.ClusterRoleBinding
-	roleBindingTr  *rbacv1.ClusterRoleBinding
+	roleBinding    *rbacv1.ClusterRoleBinding
 	serviceMonitor *monitoringv1.ServiceMonitor
 	prometheusRule *monitoringv1.PrometheusRule
 }
 
-func newMonolithReconciler(cmn *reconcilers.Instance) *flpMonolithReconciler {
-	name := name(ConfMonolith)
-	owned := monolithOwnedObjects{
-		daemonSet:      &appsv1.DaemonSet{},
+func newTransformerReconciler(cmn *reconcilers.Instance) *flpTransformerReconciler {
+	name := name(ConfKafkaTransformer)
+	owned := transfoOwnedObjects{
+		deployment:     &appsv1.Deployment{},
 		promService:    &corev1.Service{},
+		hpa:            &ascv2.HorizontalPodAutoscaler{},
 		serviceAccount: &corev1.ServiceAccount{},
 		configMap:      &corev1.ConfigMap{},
-		roleBindingIn:  &rbacv1.ClusterRoleBinding{},
-		roleBindingTr:  &rbacv1.ClusterRoleBinding{},
+		roleBinding:    &rbacv1.ClusterRoleBinding{},
 		serviceMonitor: &monitoringv1.ServiceMonitor{},
 		prometheusRule: &monitoringv1.PrometheusRule{},
 	}
-	cmn.Managed.AddManagedObject(name, owned.daemonSet)
+	cmn.Managed.AddManagedObject(name, owned.deployment)
+	cmn.Managed.AddManagedObject(name, owned.hpa)
 	cmn.Managed.AddManagedObject(name, owned.serviceAccount)
-	cmn.Managed.AddManagedObject(promServiceName(ConfMonolith), owned.promService)
-	cmn.Managed.AddManagedObject(RoleBindingMonoName(ConfKafkaIngester), owned.roleBindingIn)
-	cmn.Managed.AddManagedObject(RoleBindingMonoName(ConfKafkaTransformer), owned.roleBindingTr)
-	cmn.Managed.AddManagedObject(configMapName(ConfMonolith), owned.configMap)
+	cmn.Managed.AddManagedObject(promServiceName(ConfKafkaTransformer), owned.promService)
+	cmn.Managed.AddManagedObject(RoleBindingName(ConfKafkaTransformer), owned.roleBinding)
+	cmn.Managed.AddManagedObject(configMapName(ConfKafkaTransformer), owned.configMap)
 	if cmn.AvailableAPIs.HasSvcMonitor() {
-		cmn.Managed.AddManagedObject(serviceMonitorName(ConfMonolith), owned.serviceMonitor)
+		cmn.Managed.AddManagedObject(serviceMonitorName(ConfKafkaTransformer), owned.serviceMonitor)
 	}
 	if cmn.AvailableAPIs.HasPromRule() {
-		cmn.Managed.AddManagedObject(prometheusRuleName(ConfMonolith), owned.prometheusRule)
+		cmn.Managed.AddManagedObject(prometheusRuleName(ConfKafkaTransformer), owned.prometheusRule)
 	}
 
-	return &flpMonolithReconciler{
+	return &flpTransformerReconciler{
 		Instance: cmn,
 		owned:    owned,
 	}
 }
 
-func (r *flpMonolithReconciler) context(ctx context.Context) context.Context {
-	l := log.FromContext(ctx).WithValues(contextReconcilerName, "monolith")
+func (r *flpTransformerReconciler) context(ctx context.Context) context.Context {
+	l := log.FromContext(ctx).WithValues(contextReconcilerName, "transformer")
 	return log.IntoContext(ctx, l)
 }
 
 // cleanupNamespace cleans up old namespace
-func (r *flpMonolithReconciler) cleanupNamespace(ctx context.Context) {
+func (r *flpTransformerReconciler) cleanupNamespace(ctx context.Context) {
 	r.Managed.CleanupPreviousNamespace(ctx)
 }
 
-func (r *flpMonolithReconciler) reconcile(ctx context.Context, desired *flowslatest.FlowCollector) error {
+func (r *flpTransformerReconciler) reconcile(ctx context.Context, desired *flowslatest.FlowCollector) error {
 	// Retrieve current owned objects
 	err := r.Managed.FetchAll(ctx)
 	if err != nil {
 		return err
 	}
 
-	// Monolith only used without Kafka
-	if helper.UseKafka(&desired.Spec) {
+	// Transformer only used with Kafka
+	if !helper.UseKafka(&desired.Spec) {
 		r.Managed.TryDeleteAll(ctx)
 		return nil
 	}
 
-	builder, err := newMonolithBuilder(r.Instance, &desired.Spec)
+	builder, err := newTransfoBuilder(r.Instance, &desired.Spec)
 	if err != nil {
 		return err
 	}
@@ -107,7 +108,6 @@ func (r *flpMonolithReconciler) reconcile(ctx context.Context, desired *flowslat
 			return err
 		}
 	}
-
 	if err := r.reconcilePermissions(ctx, &builder); err != nil {
 		return err
 	}
@@ -123,20 +123,57 @@ func (r *flpMonolithReconciler) reconcile(ctx context.Context, desired *flowslat
 		return err
 	}
 
-	// Watch for Kafka exporter certificate if necessary; need to restart pods in case of cert rotation
+	// Watch for Kafka certificate if necessary; need to restart pods in case of cert rotation
+	if err = annotateKafkaCerts(ctx, r.Common, &desired.Spec.Kafka, "kafka", annotations); err != nil {
+		return err
+	}
+	// Same for Kafka exporters
 	if err = annotateKafkaExporterCerts(ctx, r.Common, desired.Spec.Exporters, annotations); err != nil {
 		return err
 	}
-
 	// Watch for monitoring caCert
 	if err = reconcileMonitoringCerts(ctx, r.Common, &desired.Spec.Processor.Metrics.Server.TLS, r.Namespace); err != nil {
 		return err
 	}
 
-	return r.reconcileDaemonSet(ctx, builder.daemonSet(annotations))
+	if err = r.reconcileDeployment(ctx, &desired.Spec.Processor, &builder, annotations); err != nil {
+		return err
+	}
+
+	return r.reconcileHPA(ctx, &desired.Spec.Processor, &builder)
 }
 
-func (r *flpMonolithReconciler) reconcilePrometheusService(ctx context.Context, builder *monolithBuilder) error {
+func (r *flpTransformerReconciler) reconcileDeployment(ctx context.Context, desiredFLP *flpSpec, builder *transfoBuilder, annotations map[string]string) error {
+	report := helper.NewChangeReport("FLP Deployment")
+	defer report.LogIfNeeded(ctx)
+
+	return reconcilers.ReconcileDeployment(
+		ctx,
+		r.Instance,
+		r.owned.deployment,
+		builder.deployment(annotations),
+		constants.FLPName,
+		helper.PtrInt32(desiredFLP.KafkaConsumerReplicas),
+		&desiredFLP.KafkaConsumerAutoscaler,
+		&report,
+	)
+}
+
+func (r *flpTransformerReconciler) reconcileHPA(ctx context.Context, desiredFLP *flpSpec, builder *transfoBuilder) error {
+	report := helper.NewChangeReport("FLP autoscaler")
+	defer report.LogIfNeeded(ctx)
+
+	return reconcilers.ReconcileHPA(
+		ctx,
+		r.Instance,
+		r.owned.hpa,
+		builder.autoScaler(),
+		&desiredFLP.KafkaConsumerAutoscaler,
+		&report,
+	)
+}
+
+func (r *flpTransformerReconciler) reconcilePrometheusService(ctx context.Context, builder *transfoBuilder) error {
 	report := helper.NewChangeReport("FLP prometheus service")
 	defer report.LogIfNeeded(ctx)
 
@@ -158,39 +195,19 @@ func (r *flpMonolithReconciler) reconcilePrometheusService(ctx context.Context, 
 	return nil
 }
 
-func (r *flpMonolithReconciler) reconcileDaemonSet(ctx context.Context, desiredDS *appsv1.DaemonSet) error {
-	report := helper.NewChangeReport("FLP DaemonSet")
-	defer report.LogIfNeeded(ctx)
-
-	return reconcilers.ReconcileDaemonSet(
-		ctx,
-		r.Instance,
-		r.owned.daemonSet,
-		desiredDS,
-		constants.FLPName,
-		&report,
-	)
-}
-
-func (r *flpMonolithReconciler) reconcilePermissions(ctx context.Context, builder *monolithBuilder) error {
+func (r *flpTransformerReconciler) reconcilePermissions(ctx context.Context, builder *transfoBuilder) error {
 	if !r.Managed.Exists(r.owned.serviceAccount) {
 		return r.CreateOwned(ctx, builder.serviceAccount())
 	} // We only configure name, update is not needed for now
 
-	cr := buildClusterRoleIngester(r.UseOpenShiftSCC)
+	cr := BuildClusterRoleTransformer()
 	if err := r.ReconcileClusterRole(ctx, cr); err != nil {
 		return err
 	}
-	cr = BuildClusterRoleTransformer()
-	if err := r.ReconcileClusterRole(ctx, cr); err != nil {
+
+	desired := builder.clusterRoleBinding()
+	if err := r.ReconcileClusterRoleBinding(ctx, desired); err != nil {
 		return err
-	}
-	// Monolith uses ingester + transformer cluster roles
-	for _, kind := range []ConfKind{ConfKafkaIngester, ConfKafkaTransformer} {
-		desired := builder.clusterRoleBinding(kind)
-		if err := r.ReconcileClusterRoleBinding(ctx, desired); err != nil {
-			return err
-		}
 	}
 
 	return reconcileLokiRoles(ctx, r.Common, &builder.generic)
